+++ backend/app/schema/savings.py
@@ -4,14 +4,15 @@ Schema models
 from datetime import datetime
 from typing import Optional
 from pydantic import BaseModel, ConfigDict, field_validator
+from decimal import Decimal
 
 
 class SavingsCreate(BaseModel):
     """
     Schema for create
     """
-    amount: float
-    current_amount: float = 0.0
+    amount: Decimal
+    current_amount: Decimal = Decimal(0.0)
     target_date: Optional[datetime] = None
     duration_months: Optional[int] = None
     description: Optional[str] = None
@@ -36,8 +37,8 @@ class SavingsUpdate(BaseModel):
     """
     Schema for update
     """
-    amount: Optional[float]
-    current_amount: Optional[float]
+    amount: Optional[Decimal]
+    current_amount: Optional[Decimal]
     target_date: Optional[datetime]
     duration_months: Optional[int]
     description: Optional[str]
